Hash64 accepts the length as a signed integer

+++ src/Farmhash.Sharp.Benchmarks/Program.cs
@@ -43,7 +43,7 @@ namespace Farmhash.Sharp.Benchmarks
             Console.WriteLine();
             Console.WriteLine("Iterations per millisecond 64bit");
             Console.WriteLine("Name\tSmallest\tSmaller\tSmall\tMedium\tLarge\tLarge st");
-            ProfileSuite("Farmhash", 10000, bytes => Farmhash.Hash64(bytes, (ulong) bytes.LongLength));
+            ProfileSuite("Farmhash", 10000, bytes => Farmhash.Hash64(bytes, bytes.LongLength));
             ProfileSuite("xxHashSharp", 10000, bytes => xxHash.CalculateHash(bytes));
             ProfileSuite("CityHash.Net", 1000, str => CityHash.CityHash.CityHash64(str));
             ProfileSuite("HashFunction CityHash", 10000, (byte[] bytes) => hcity64.ComputeHash(bytes));
